Duel Links Scraping Project
In this project I scraped data from the https://www.duellinksmeta.com/ website. The aim of this project is to show a count of a specific card for each competitive deck made by a professional player. During this project, I scraped the name of the cards of each deck, the URL address of each deck, the name of the author of the sample deck, and then merged them together into a sorted list that was ordered by the count of cards. With this assortment, you could visibly and easily notice which card is used the most from all of the competitive decks, and therefore to recognize the game meta as of the moment of scraping.